requirements.txt
psycopg2-binary
kagglehub
pandas
sqlalchemy
shutil